feat: patient registration swagger + get location
