+++ packages/service/core/openapi.yaml
@@ -22,7 +22,7 @@ paths:
     post:
       summary: Medical practitoner registration
       description: Register user as medical practitoner
-      operationId: createUser
+      operationId: createMedicalPractitonerUser
       responses:
         '200':
           description: User created successfully
@@ -48,6 +48,36 @@ paths:
           application/x-www-form-urlencoded:
             schema:
               $ref: '#/components/schemas/MedicalPractitoner'
+  /user/register/individual:
+    post:
+      summary: Patient registration
+      description: Register user as patient
+      operationId: createIndividualUser
+      responses:
+        '200':
+          description: User created successfully
+          content:
+            application/xml:
+              schema:
+                $ref: '#/components/schemas/Individual'
+            application/json:
+              schema:
+                $ref: '#/components/schemas/Individual'
+        '405':
+          description: Invalid input
+      requestBody:
+        description: Create a new patient user
+        required: true
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/Individual'
+          application/xml:
+            schema:
+              $ref: '#/components/schemas/Individual'
+          application/x-www-form-urlencoded:
+            schema:
+              $ref: '#/components/schemas/Individual'
 components:
   schemas:
     MedicalPractitoner:
@@ -92,6 +122,56 @@ components:
           example: false
           description: Email Status
       xml:
-        name: user
-      type: object      
+        name: medical_practitoner
+      type: object
+    Individual:
+      properties:
+        id:
+          type: integer
+          format: int64
+          example: 11
+        name:
+          type: string
+          example: Richard Branson
+        email:
+          type: string
+          example: richard@email.com
+        country:
+          type: string
+          example: Australia
+        location:
+          type: string
+          example: "12.00,43.02"
+        timezone:
+          type: string
+          example: "Sydney/Australia"
+        preferredSpecialist:
+          type: string
+          example: "Dermatology"
+        languages:
+          type: array
+          example: ["Tamil", "English"]
+          items:
+            type: string
+        preferredConsultation:
+          type: array
+          example: ["group", "one-to-one"]
+          items:
+            type: string
+        preferredConsultationDate:
+          type: string
+          example: "2022-05-16T10:19:00.000Z"
+        phoneNumber:
+          type: string
+          example: "+61406210236"
+        emailVerified:
+          type: boolean
+          example: false
+          description: Email Status
+        additionalMessage:
+          type: string
+          example: "Additional consultation message"
+      xml:
+        name: individual
+      type: object       
 tags: []
