--- packages/service/core/src/db/models/Patient.ts
@@ -1,10 +1,0 @@
-export const schema = {
-  name: String,
-  phoneNumber: String,
-  timeZone: String,
-  languages: [String],
-  completedAppointments: [String],
-  upcomingAppointments: [String]
-}
-
-export const model = schema
