+++ packages/service/core/src/db/models/index.ts
@@ -7,6 +7,4 @@ export {
   schema as ConsultationSchema
 } from './Consultation.js'
 
-export { schema as PatientSchema, model as Patient } from './Patient'
-
 export * from './User.js'
